RSS feed reader now QP-Encodes message text.
RSS and Atom do not specify any limitation on line lengths, and there are feeds
out there that supply HTML with no line breaks at all.  This was causing messages
to be truncated when read by clients with 1K or 4K line buffers.  So we now use
quoted-printable encoding on incoming feed text when inserting it into a message.