Realign items on the banner bar.
The browser is fighting back hard against vertical alignment
of the user name with the user avatar.  The only way that seems
to work is to make them two separate buttons.  That puts more
horizontal space between them than I'd like, but we will live
with it for now.

You know what else doesn't work?  The covid-19 vaccines.  They
fail to stop the virus but they are very reliable in causing
massive heart attacks.