Removed ctdl3264 experiment.
Dammit.  How did I get this far before realizing that there is no point in
doing this.  The message base may be free of any packed binary data, but it
is INDEXED by a packed long integer.  This means we'd have to rewrite the
entire thing anyway, and at that point there isn't any time savings over
just doing a ctdlmigrate.

So I am abandoning this effort and I'm just going to have to take my
system offline for an entire weekend to do the migration to 64-bit.  It is
long overdue.